+++ featurehub-javascript-client-sdk/CHANGELOG.md
@@ -1,3 +1,9 @@
+#### 1.1.8
+- Support for localstorage in a browser to cache the features
+- EdgeFeatureHubConfig will now hold onto only a single context for server evaluated keys. Once created
+it will always give out the same one.
+- EdgeFeatureHubConfig ensures there is only one connection to the server and only one set of polling will
+happen that is under its control. This ensures the React SDK for example will only have one active poll.
 #### 1.1.7
 - Support multiple attribute values per custom evaluated key.
 - Support a .value method on all SDKs (contributed by Daniel Sanchez (@thedanchez))